+++ app/components/helper/custom-success-icon.jsx
@@ -1,0 +1,18 @@
+import React from "react";  
+
+const CustomSuccessIcon = () => {
+  return (
+    <svg viewBox="0 0 24 24" width="24" height="24">
+      <circle cx="12" cy="12" r="12" fill="oklch(60.6% 0.25 292.717)" />
+      <path
+        d="M7 12l4 4 6-8"
+        stroke="#16f2b3"
+        strokeWidth="2"
+        strokeLinecap="round"
+        fill="none"
+      />
+    </svg>
+  );
+}
+
+export default CustomSuccessIcon
