+++ app/components/homepage/contact/contact-form.jsx
@@ -1,10 +1,13 @@
 "use client";
 // @flow strict
-import { isValidEmail } from "@/utils/check-email";
-import axios from "axios";
 import { useState } from "react";
 import { TbMailForward } from "react-icons/tb";
 import { toast } from "react-toastify";
+import axios from "axios";
+
+import CustomErrorIcon from "../../helper/custom-error-icon";
+import CustomSuccessIcon from "../../helper/custom-success-icon";
+import { isValidEmail } from "@/utils/check-email";
 
 function ContactForm() {
   const [error, setError] = useState({ email: false, required: false });
@@ -14,35 +17,6 @@ function ContactForm() {
     email: "",
     message: "",
   });
-  
-  const CustomErrorIcon = () => {
-    return (
-      <svg viewBox="0 0 24 24" width="24" height="24">
-        <circle cx="12" cy="12" r="12" fill="#121212" />
-        <path
-          d="M8 8l8 8M16 8l-8 8"
-          stroke="#e74c3c"
-          strokeWidth="2"
-          strokeLinecap="round"
-        />
-      </svg>
-    );
-  }
-  
-  const CustomSuccessIcon = () => {
-    return (
-      <svg viewBox="0 0 24 24" width="24" height="24">
-        <circle cx="12" cy="12" r="12" fill="oklch(60.6% 0.25 292.717)" />
-        <path
-          d="M7 12l4 4 6-8"
-          stroke="#16f2b3"
-          strokeWidth="2"
-          strokeLinecap="round"
-          fill="none"
-        />
-      </svg>
-    );
-  }
 
   const checkRequired = () => {
     if (userInput.email && userInput.message && userInput.name) {
